chore: refactor code (see description)
- updated wrapped-sample-document.json with the new OA v3 wrapper
- fixed some imports in digest
- fixed wrong schema import in /v2/wrap.ts (we accidentally used v3 schema in v2)
- since digest has a bit of diff between v2/v3, we moved it into its respective version folders
