+++ src/v3/digest/digest.test.ts
@@ -1,0 +1,7 @@
+import { digestDocument as digestDocumentV3 } from "./digest";
+
+describe.skip("digest v3.0", () => {
+  describe("placeholder for future tests", () => {
+
+  });
+});
