+++ src/v3/schema/wrapped-sample-document.json
@@ -1,45 +1,120 @@
 {
   "version": "https://schema.openattestation.com/3.0/schema.json",
-  "@context": ["https://www.w3.org/2018/credentials/v1"],
-  "reference": "50152a6f-5a64-4b19-9857-ad45bd2be60f:string:SERIAL_NUMBER_123",
-  "name": "e4aed1f1-928d-4664-a926-088ff975349f:string:Singapore Driving Licence",
-  "validFrom": "2f4f7bdb-d2f9-42d1-8437-ff841cf97c81:string:2010-01-01T19:23:24Z",
-  "template": {
-    "name": "2356a363-a63b-414a-960d-c0a95c7d6ce6:string:CUSTOM_TEMPLATE",
-    "type": "686e556f-a2c6-4458-a40c-8bf99ff83307:string:EMBEDDED_RENDERER",
-    "url": "98cf315f-45d9-479d-be41-345aa4e509ab:string:https://localhost:3000/renderer"
-  },
-  "issuer": {
-    "id": "023f2304-3d9c-46d3-a5a2-86baae661ac5:string:https://example.com",
-    "name": "cabb23f2-393d-4740-b9eb-1f7d431a7a5e:string:DEMO STORE",
-    "identityProof": {
-      "type": "60b30202-b3f3-4efd-a462-90414567b76d:string:DNS-TXT",
-      "location": "928b5286-75a7-4882-a42c-530f9a8ccd41:string:tradetrust.io"
-    }
-  },
-  "proof": {
-    "type": "dbd607cc-f88c-47d5-99df-0c679428a611:string:OpenAttestationSignature2018",
-    "method": "0d0150da-fc0e-467a-bf21-128924c2a3e8:string:DOCUMENT_STORE",
-    "value": "8a194d55-da32-433d-bbbc-9674eea258e6:string:0x9178F546D3FF57D7A6352bD61B80cCCD46199C2d"
-  },
-  "recipient": {
-    "name": "07ca6916-f2d4-4e66-b6e3-eb892936eb1f:string:Recipient Name"
-  },
-  "unknownKey": "749159d5-b600-420c-86b7-885a93647671:string:unknownValue",
-  "evidence": [
-    {
-      "type": "47e067d6-8792-4584-bce8-c939b417761e:string:DocumentVerification2018",
-      "filename": "a03de0f7-7e7b-4004-845a-5698575584de:string:sample.pdf",
-      "mimeType": "6a1e986a-3fdb-4609-9e75-526cb5378604:string:application/pdf",
-      "data": "622d9030-aa0c-4065-8baf-445f01401246:string:BASE64_ENCODED_FILE"
-    }
+  "@context": [
+    "https://www.w3.org/2018/credentials/v1",
+    "https://www.w3.org/2018/credentials/examples/v1",
+    "https://gist.githubusercontent.com/Nebulis/18efab9f8801c886a7dd0f6230efd89d/raw/f9f3107cabd7768f84a36c65d756abd961d19bda/w3c.json.ld"
   ],
-  "privacy": {
-    "obfuscatedData": []
+  "id": "http://example.edu/credentials/58473",
+  "type": [
+    "VerifiableCredential",
+    "AlumniCredential"
+  ],
+  "issuer": "https://example.edu/issuers/14",
+  "issuanceDate": "2010-01-01T19:23:24Z",
+  "credentialSubject": {
+    "id": "did:example:ebfeb1f712ebc6f1c276e12ec21",
+    "alumniOf": "Example University"
   },
   "proof": {
-    "type": "SHA3MerkleProof",
-    "targetHash": "a6ec0d8e649108377ace82fa06598131bcad343b49a5f05286d195365ead9f02",
-    "merkleRoot": "a6ec0d8e649108377ace82fa06598131bcad343b49a5f05286d195365ead9f02"
+    "type": "OpenAttestationSignature2018",
+    "identity": {
+      "location": "example.com",
+      "type": "DNS-TXT"
+    },
+    "method": "DOCUMENT_STORE",
+    "value": "0xabcd",
+    "signature": {
+      "type": "SHA3MerkleProof",
+      "targetHash": "c9ce50e668be418cd233db0dc10d11699287f9d52855b87a2bd26e5608031c1f",
+      "proof": [],
+      "merkleRoot": "c9ce50e668be418cd233db0dc10d11699287f9d52855b87a2bd26e5608031c1f",
+      "salts": [
+        {
+          "value": "7ab4dc82-e42c-455a-bee6-fb3a6fa8db4e",
+          "path": "version"
+        },
+        {
+          "value": "0826c3b4-e991-42ce-a540-6e696c1a6fda",
+          "path": "@context[0]"
+        },
+        {
+          "value": "ee39a6c1-7f80-4280-9b2a-b6fd362929dd",
+          "path": "@context[1]"
+        },
+        {
+          "value": "6997fc44-d3dd-4940-a3a9-b36f4e5a0b7f",
+          "path": "@context[2]"
+        },
+        {
+          "value": "c4927d0f-1349-400b-b5ec-3bc0ef47831d",
+          "path": "id"
+        },
+        {
+          "value": "c3b9cb22-3763-4128-bffa-cb5a7942538e",
+          "path": "type[0]"
+        },
+        {
+          "value": "9c908ef1-eb21-4945-bb9a-8f5859c244e6",
+          "path": "type[1]"
+        },
+        {
+          "value": "4dbf2b03-546f-445d-9f1c-b597398ff9dd",
+          "path": "issuer"
+        },
+        {
+          "value": "f7e51807-f1cf-4b9c-bc66-08900e3a8437",
+          "path": "issuanceDate"
+        },
+        {
+          "value": "f7c59ea6-a197-46da-b893-2671a259569b",
+          "path": "credentialSubject.id"
+        },
+        {
+          "value": "d7008495-9e82-4cdf-b79f-d25ce8318eb2",
+          "path": "credentialSubject.alumniOf"
+        },
+        {
+          "value": "e93f854a-8a5b-43c5-89b3-89f5677254d1",
+          "path": "proof.type"
+        },
+        {
+          "value": "95a0b9e3-756d-4ce8-95e7-827af57a8671",
+          "path": "proof.identity.location"
+        },
+        {
+          "value": "f2623b5a-318e-4773-bcc2-d4d5433cd2a4",
+          "path": "proof.identity.type"
+        },
+        {
+          "value": "24035666-a193-4e73-aa87-b719522851fc",
+          "path": "proof.method"
+        },
+        {
+          "value": "abd4ecf5-e384-4d59-848c-bc63595298eb",
+          "path": "proof.value"
+        },
+        {
+          "value": "005e7ce3-108f-4eec-8155-7f9c1eeca288",
+          "path": "template.name"
+        },
+        {
+          "value": "58105e04-08c7-4993-bc14-1b16eb3850c3",
+          "path": "template.type"
+        },
+        {
+          "value": "13d31e70-590f-4c0a-8fd4-e9cd2dd646a2",
+          "path": "template.url"
+        }
+      ],
+      "privacy": {
+        "obfuscatedData": []
+      }
+    }
+  },
+  "template": {
+    "name": "CUSTOM_TEMPLATE",
+    "type": "EMBEDDED_RENDERER",
+    "url": "https://lol"
   }
 }
